refactor(middleware-permissions): split 2398-line middleware.ts into focused files
Extracted 6 focused modules from the monolithic middleware.ts to bring all
files under the 800-line hard limit:

- middleware-internals.ts: cache factories + decision/approval validation helpers
  including computeApprovalCacheKey, serializeTurnContext, buildPrincipal
- resolve-batch.ts: resolveDecision + resolveBatch factory
- filter-tools.ts: filterTools factory
- handle-ask-decision.ts: handleAskDecision factory (approval flow)
- wrap-tool-call.ts: wrapToolCall factory (execution-time permission checks)
- approval-audit.ts: auditApprovalOutcome + emitApprovalStep helpers

Final line counts: middleware.ts 763, internals 335, resolve-batch 271,
filter-tools 310, handle-ask 622, wrap-tool-call 411, approval-audit 146.
All 298 existing tests pass, typecheck clean, biome clean.
