+++ packages/security/middleware-permissions/src/middleware.ts
@@ -13,7 +13,6 @@ import { canonicalPrefix, classifyCommand, UNSAFE_PREFIX } from "@koi/bash-class
 import type { AuditEntry, AuditSink } from "@koi/core";
 import type { JsonObject } from "@koi/core/common";
 import type {
-  ApprovalHandler,
   CapabilityFragment,
   KoiMiddleware,
   ModelHandler,
@@ -28,20 +27,12 @@ import type {
 } from "@koi/core/middleware";
 import type { PermissionDecision, PermissionQuery } from "@koi/core/permission-backend";
 import type { RichTrajectoryStep } from "@koi/core/rich-trajectory";
-import {
-  type CircuitBreaker,
-  createCircuitBreaker,
-  KoiRuntimeError,
-  swallowError,
-} from "@koi/errors";
+import { type CircuitBreaker, createCircuitBreaker, swallowError } from "@koi/errors";
 import { computeStringHash } from "@koi/hash";
+import { createApprovalAudit } from "./approval-audit.js";
 // fnv1a no longer used for cache keys (collision-unsafe for security decisions)
 import { isDefaultDeny } from "./classifier.js";
-import type {
-  ApprovalCacheConfig,
-  PermissionCacheConfig,
-  PermissionsMiddlewareConfig,
-} from "./config.js";
+import type { PermissionsMiddlewareConfig } from "./config.js";
 import {
   DEFAULT_APPROVAL_CACHE_MAX_ENTRIES,
   DEFAULT_APPROVAL_CACHE_TTL_MS,
@@ -49,329 +40,32 @@ import {
   DEFAULT_CACHE_CONFIG,
   DEFAULT_DENIAL_ESCALATION_THRESHOLD,
   DEFAULT_DENIAL_ESCALATION_WINDOW_MS,
-  DEFAULT_SOFT_DENY_PER_TURN_CAP,
 } from "./config.js";
 import { createDenialTracker, type DenialTracker } from "./denial-tracker.js";
+import { createFilterTools } from "./filter-tools.js";
+import { createHandleAskDecision } from "./handle-ask-decision.js";
+import {
+  buildPrincipal,
+  computeApprovalCacheKey,
+  createApprovalCache,
+  createDecisionCache,
+  decisionCacheKey,
+  serializeTurnContext,
+  VALID_ALWAYS_ALLOW_SCOPES,
+  VALID_APPROVAL_KINDS,
+  VALID_EFFECTS,
+} from "./middleware-internals.js";
+import { createBatchResolver } from "./resolve-batch.js";
 import { createSoftDenyLog, type SoftDenyLog } from "./soft-deny-log.js";
 import { createTurnSoftDenyCounter, type TurnSoftDenyCounter } from "./turn-soft-deny-counter.js";
+import { createWrapToolCall } from "./wrap-tool-call.js";
 
 // ---------------------------------------------------------------------------
-// Internal cache types
+// Re-exported internals (for tests / consumers that import these symbols)
 // ---------------------------------------------------------------------------
 
-interface DecisionCacheEntry {
-  readonly decision: PermissionDecision;
-  readonly expiresAt: number;
-}
-
-interface ApprovalCacheEntry {
-  readonly cachedAt: number;
-}
-
-// ---------------------------------------------------------------------------
-// Decision cache
-// ---------------------------------------------------------------------------
-
-function createDecisionCache(
-  config: PermissionCacheConfig,
-  clock: () => number,
-): {
-  readonly get: (key: string) => PermissionDecision | undefined;
-  readonly set: (key: string, decision: PermissionDecision) => void;
-  readonly clear: () => void;
-} {
-  const maxEntries = config.maxEntries ?? DEFAULT_CACHE_CONFIG.maxEntries;
-  const allowTtl = config.allowTtlMs ?? DEFAULT_CACHE_CONFIG.allowTtlMs;
-  const denyTtl = config.denyTtlMs ?? DEFAULT_CACHE_CONFIG.denyTtlMs;
-  // String keys: collision-safe for security-sensitive authorization decisions
-  const entries = new Map<string, DecisionCacheEntry>();
-
-  return {
-    get(key) {
-      const entry = entries.get(key);
-      if (entry === undefined) return undefined;
-      // expiresAt === Infinity means no expiry (ttl was 0)
-      if (entry.expiresAt !== Infinity && clock() >= entry.expiresAt) {
-        entries.delete(key);
-        return undefined;
-      }
-      // LRU: move to end
-      entries.delete(key);
-      entries.set(key, entry);
-      return entry.decision;
-    },
-
-    set(key, decision) {
-      // Never cache "ask" decisions
-      if (decision.effect === "ask") return;
-
-      const ttl = decision.effect === "allow" ? allowTtl : denyTtl;
-      // Evict oldest if full
-      if (entries.size >= maxEntries) {
-        const oldest = entries.keys().next().value;
-        if (oldest !== undefined) entries.delete(oldest);
-      }
-      // ttl === 0 means no expiry (permanent cache until eviction or clear)
-      entries.set(key, { decision, expiresAt: ttl === 0 ? Infinity : clock() + ttl });
-    },
-
-    clear() {
-      entries.clear();
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Approval cache
-// ---------------------------------------------------------------------------
-
-function createApprovalCache(
-  config: ApprovalCacheConfig,
-  clock: () => number,
-): {
-  readonly has: (key: string) => boolean;
-  readonly set: (key: string) => void;
-  readonly clear: () => void;
-} {
-  const maxEntries = config.maxEntries ?? DEFAULT_APPROVAL_CACHE_MAX_ENTRIES;
-  const ttlMs = config.ttlMs ?? DEFAULT_APPROVAL_CACHE_TTL_MS;
-  // String keys: collision-safe for security-sensitive approval decisions
-  const entries = new Map<string, ApprovalCacheEntry>();
-
-  return {
-    has(key) {
-      const entry = entries.get(key);
-      if (entry === undefined) return false;
-      if (ttlMs > 0 && clock() - entry.cachedAt >= ttlMs) {
-        entries.delete(key);
-        return false;
-      }
-      // LRU: move to end
-      entries.delete(key);
-      entries.set(key, entry);
-      return true;
-    },
-
-    set(key) {
-      if (entries.size >= maxEntries) {
-        const oldest = entries.keys().next().value;
-        if (oldest !== undefined) entries.delete(oldest);
-      }
-      entries.set(key, { cachedAt: clock() });
-    },
-
-    clear() {
-      entries.clear();
-    },
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Decision validation (fail-closed on malformed backend responses)
-// ---------------------------------------------------------------------------
-
-const VALID_EFFECTS = new Set(["allow", "deny", "ask"]);
-
-/** Symbol to tag fail-closed denials so escalation can exclude them. */
-const IS_FAIL_CLOSED: unique symbol = Symbol.for("@koi/middleware-permissions/fail-closed");
-
-const FAIL_CLOSED_DENY: PermissionDecision = Object.freeze({
-  effect: "deny",
-  reason: "Malformed backend decision — failing closed",
-  [IS_FAIL_CLOSED]: true,
-} as PermissionDecision);
-
-function isFailClosed(decision: PermissionDecision): boolean {
-  return (decision as Record<symbol, unknown>)[IS_FAIL_CLOSED] === true;
-}
-
-function failClosedDeny(reason: string): PermissionDecision {
-  return Object.freeze({
-    effect: "deny",
-    reason,
-    [IS_FAIL_CLOSED]: true,
-  } as PermissionDecision);
-}
-
-/** Symbol to tag escalation-generated denials so they don't self-sustain. */
-const IS_ESCALATED: unique symbol = Symbol.for("@koi/middleware-permissions/escalated");
-
-function isEscalated(decision: PermissionDecision): boolean {
-  return (decision as Record<symbol, unknown>)[IS_ESCALATED] === true;
-}
-
-/** Symbol to tag cached deny replays so they don't inflate escalation counts. */
-const IS_CACHED: unique symbol = Symbol.for("@koi/middleware-permissions/cached");
-
-function isCached(decision: PermissionDecision): boolean {
-  return (decision as Record<symbol, unknown>)[IS_CACHED] === true;
-}
-
-function tagCached(decision: PermissionDecision): PermissionDecision {
-  if (decision.effect !== "deny") return decision;
-  return { ...decision, [IS_CACHED]: true } as PermissionDecision;
-}
-
-/** Determine denial source for tracker recording. Only "policy" counts toward escalation. */
-/** Safely serialize a value to a JSON preview string, truncated to maxLen. */
-function safePreviewJson(value: unknown, maxLen: number): string {
-  try {
-    const s = JSON.stringify(value);
-    return s.length <= maxLen ? s : `${s.slice(0, maxLen)}…`;
-  } catch {
-    return "[unserializable]";
-  }
-}
-
-function denialSource(
-  decision: PermissionDecision,
-): "policy" | "backend-error" | "escalation" | "approval" {
-  if (isFailClosed(decision)) return "backend-error";
-  if (isEscalated(decision)) return "escalation";
-  if (isCached(decision)) return "escalation"; // cached replays must not inflate escalation
-  return "policy";
-}
-
-/**
- * Validate a backend decision at the trust boundary. Malformed or
- * unexpected shapes are converted to deny (fail-closed) rather than
- * silently falling through to allow.
- */
-function validateDecision(raw: unknown): PermissionDecision {
-  if (raw === null || typeof raw !== "object") return FAIL_CLOSED_DENY;
-  const obj = raw as Record<string, unknown>;
-  if (!VALID_EFFECTS.has(obj.effect as string)) return FAIL_CLOSED_DENY;
-  if (obj.effect === "deny" || obj.effect === "ask") {
-    if (typeof obj.reason !== "string") return FAIL_CLOSED_DENY;
-  }
-  return raw as PermissionDecision;
-}
-
-const VALID_APPROVAL_KINDS = new Set(["allow", "always-allow", "deny", "modify"]);
-const VALID_ALWAYS_ALLOW_SCOPES = new Set(["session", "always"]);
-
-/**
- * Validate an approval handler response at the trust boundary.
- * Returns the validated decision or undefined if malformed (caller
- * should fail closed).
- */
-function validateApprovalDecision(
-  raw: unknown,
-):
-  | { readonly kind: "allow" }
-  | { readonly kind: "always-allow"; readonly scope: "session" | "always" }
-  | { readonly kind: "deny"; readonly reason: string }
-  | { readonly kind: "modify"; readonly updatedInput: Record<string, unknown> }
-  | undefined {
-  if (raw === null || typeof raw !== "object") return undefined;
-  const obj = raw as Record<string, unknown>;
-  if (!VALID_APPROVAL_KINDS.has(obj.kind as string)) return undefined;
-  if (obj.kind === "deny") {
-    if (typeof obj.reason !== "string") return undefined;
-    return { kind: "deny", reason: obj.reason };
-  }
-  if (obj.kind === "modify") {
-    // Must be a plain JSON object — reject null, arrays, and non-objects
-    if (
-      obj.updatedInput === null ||
-      typeof obj.updatedInput !== "object" ||
-      Array.isArray(obj.updatedInput)
-    ) {
-      return undefined;
-    }
-    return { kind: "modify", updatedInput: obj.updatedInput as Record<string, unknown> };
-  }
-  if (obj.kind === "always-allow") {
-    const scope = obj.scope as string;
-    if (!VALID_ALWAYS_ALLOW_SCOPES.has(scope)) return undefined;
-    return { kind: "always-allow", scope: scope as "session" | "always" };
-  }
-  return { kind: "allow" };
-}
-
-// ---------------------------------------------------------------------------
-// Cache key helpers
-// ---------------------------------------------------------------------------
-
-/**
- * Safe JSON serialization — returns undefined on non-serializable values
- * (cyclic objects, BigInt, etc.) instead of throwing.
- */
-function safeStringify(value: unknown): string | undefined {
-  try {
-    return JSON.stringify(value);
-  } catch {
-    return undefined;
-  }
-}
-
-/** Full serialized key — collision-safe for security-sensitive cache lookups. */
-function decisionCacheKey(query: PermissionQuery): string | undefined {
-  const ctx = query.context !== undefined ? safeStringify(query.context) : "";
-  if (ctx === undefined) return undefined;
-  return `${query.principal}\0${query.action}\0${query.resource}\0${ctx}`;
-}
-
-/** Full serialized key — collision-safe for approval cache lookups. */
-function computeApprovalCacheKey(
-  backendFingerprint: string,
-  sessionId: string,
-  userId: string,
-  agentId: string,
-  toolId: string,
-  input: unknown,
-  context: string | undefined,
-  requestMeta: unknown,
-  approvalReason: string,
-  // For bash-like tools, includes the derived bash grant key so a
-  // cached one-off allow in one execution context cannot replay in
-  // another. For non-bash tools, `grantKey === toolId` and this has
-  // no effect on cache key identity.
-  grantKey: string,
-): string | undefined {
-  if (context === undefined) return undefined;
-  const sorted = safeSerializeInput(input);
-  if (sorted === undefined) return undefined;
-  const reqMeta = requestMeta !== undefined ? safeStringify(requestMeta) : "";
-  if (reqMeta === undefined) return undefined;
-  return `${backendFingerprint}\0${sessionId}\0${userId}\0${agentId}\0${toolId}\0${grantKey}\0${sorted}\0${context}\0${reqMeta}\0${approvalReason}`;
-}
-
-/** Serialize turn-scoped context for inclusion in approval cache keys. */
-/** Returns undefined when metadata is not serializable — caller must skip caching. */
-function serializeTurnContext(ctx: TurnContext): string | undefined {
-  const sessionMeta = ctx.session.metadata;
-  const turnMeta = ctx.metadata;
-  const hasSession = Object.keys(sessionMeta).length > 0;
-  const hasTurn = Object.keys(turnMeta).length > 0;
-  if (!hasSession && !hasTurn) return "";
-  const combined = { ...(hasSession ? { _s: sessionMeta } : {}), ...(hasTurn ? turnMeta : {}) };
-  return safeStringify(combined);
-}
-
-function safeSerializeInput(value: unknown): string | undefined {
-  if (value === null || typeof value !== "object") {
-    return safeStringify(value);
-  }
-  const obj = value as Record<string, unknown>;
-  const sorted = Object.keys(obj)
-    .sort()
-    .reduce<Record<string, unknown>>((acc, key) => {
-      acc[key] = obj[key];
-      return acc;
-    }, {});
-  return safeStringify(sorted);
-}
-
-/**
- * Build an unambiguous principal string from structured identity fields.
- * Uses JSON array encoding to prevent separator collisions — e.g. an
- * agentId containing ":" cannot produce the same principal as a
- * different agent/user/session tuple.
- */
-function buildPrincipal(agentId: string, userId: string, sessionId: string): string {
-  return JSON.stringify([agentId, userId, sessionId]);
-}
+// VALID_EFFECTS, VALID_APPROVAL_KINDS, VALID_ALWAYS_ALLOW_SCOPES are used in tests
+export { VALID_ALWAYS_ALLOW_SCOPES, VALID_APPROVAL_KINDS, VALID_EFFECTS };
 
 // ---------------------------------------------------------------------------
 // Factory
@@ -660,29 +354,6 @@ export function createPermissionsMiddleware(
   // -----------------------------------------------------------------------
 
   /**
-   * Derive the canonical permission resource key for a tool call. When
-   * `resolveBashCommand` is configured and returns a non-empty command
-   * string, the key is `<toolId>:<prefix>` (e.g. `bash:git push`). Returns
-   * the plain tool name otherwise.
-   */
-  /**
-   * Shared derivation for both policy and grant keys. Returns `null`
-   * when bash enrichment is not applicable (no resolver, empty
-   * command, or unresolvable prefix) so callers can fall back to the
-   * plain tool id — matches storage semantics for unenriched grants.
-   */
-  /**
-   * Combine the plain-tool decision with the enriched-resource decision.
-   *
-   * Default-deny on EITHER side is treated as "no opinion" — the
-   * operator did not write a rule at that granularity, so the other
-   * side's explicit rule should take effect. Preserves legacy plain
-   * rules like `allow: ["bash"]` / `ask: ["group:runtime"]` when no
-   * enriched rule matches, and still enforces fine-grained prefix
-   * rules when the plain side is silent. When both are explicit,
-   * deny > ask > allow.
-   */
-  /**
    * Backend-tolerant default-deny check. Recognizes both the
    * built-in pattern backend's internal symbol marker and a public
    * `default: true` field that custom backends can set on their
@@ -846,574 +517,75 @@ export function createPermissionsMiddleware(
     });
   }
 
-  /** Validated approval decision — structural type matching validateApprovalDecision return. */
-  type ValidatedApproval =
-    | { readonly kind: "allow" }
-    | { readonly kind: "always-allow"; readonly scope: "session" | "always" }
-    | { readonly kind: "deny"; readonly reason: string }
-    | { readonly kind: "modify"; readonly updatedInput: Record<string, unknown> };
+  // Wire up approval audit helpers via factory
+  const { auditApprovalOutcome, emitApprovalStep } = createApprovalAudit({
+    clock,
+    approvalSink,
+  });
 
-  /** Audit an approval outcome. Called both after user responds and on fast-path replay. */
-  function auditApprovalOutcome(
-    ctx: TurnContext,
-    resource: string,
-    approval: ValidatedApproval,
-    originalInput: JsonObject,
-    durationMs: number,
-    sink: AuditSink,
-    coalesced = false,
-    remembered = false,
-  ): void {
-    // "remembered" = fast-path replay (persistent or session grant matched).
-    // "granted" / "denied" = user responded to a prompt.
-    const permissionEvent = remembered
-      ? "remembered"
-      : approval.kind === "deny"
-        ? "denied"
-        : "granted";
-    const meta: Record<string, unknown> = {
-      permissionCheck: true,
-      permissionEvent,
-      phase: "approval_outcome",
-      resource,
-      approvalDecision: approval.kind,
-      userId: ctx.session.userId ?? "__anonymous__",
-      ...(coalesced ? { coalesced: true } : {}),
-    };
-    if (approval.kind === "deny") {
-      meta.denyReason = approval.reason;
-    }
-    if (approval.kind === "modify") {
-      // Log key names only — raw inputs may contain secrets or sensitive data.
-      // Full payload capture requires a dedicated secure-audit mode.
-      meta.originalInputKeys = Object.keys(originalInput).sort();
-      meta.modifiedInputKeys = Object.keys(approval.updatedInput).sort();
-      meta.inputModified = true;
-    }
-    if (approval.kind === "always-allow") {
-      meta.scope = approval.scope;
-    }
-    const entry: AuditEntry = {
-      schema_version: 2,
-      timestamp: clock(),
-      sessionId: ctx.session.sessionId as string,
-      agentId: ctx.session.agentId,
-      turnIndex: ctx.turnIndex,
-      kind: "permission_decision",
-      durationMs,
-      metadata: meta as JsonObject,
-    };
-    void sink.log(entry).catch((e: unknown) => {
-      swallowError(e, { package: PKG, operation: "audit-approval" });
-    });
-  }
+  // Wire up batch resolver (resolveDecision + resolveBatch) via factory
+  const { resolveDecision, resolveBatch } = createBatchResolver({
+    backend,
+    cb,
+    getDecisionCache,
+    escalationEnabled,
+    escalationThreshold,
+    escalationWindowMs,
+    clock,
+    getTracker,
+    decisionCacheKey,
+  });
 
-  /** Emit a source:"user" trajectory step for an approval decision. */
-  function emitApprovalStep(
-    ctx: TurnContext,
-    toolId: string,
-    approval: ValidatedApproval,
-    originalInput: JsonObject,
-    startMs: number,
-    coalesced = false,
-  ): void {
-    if (originalSink === undefined && runtimeSinks.length === 0) return;
-    const meta: Record<string, unknown> = {
-      approvalDecision: approval.kind,
-      userId: ctx.session.userId ?? "__anonymous__",
-      ...(coalesced ? { coalesced: true } : {}),
-    };
-    if (approval.kind === "modify") {
-      meta.inputModified = true;
-      meta.originalInputKeys = Object.keys(originalInput).sort();
-      meta.modifiedInputKeys = Object.keys(approval.updatedInput).sort();
-    }
-    if (approval.kind === "deny") {
-      meta.denyReason = approval.reason;
-    }
-    if (approval.kind === "always-allow") {
-      meta.scope = approval.scope;
-    }
-    const step: RichTrajectoryStep = {
-      stepIndex: -1,
-      timestamp: startMs,
-      source: "user",
-      kind: "tool_call",
-      identifier: toolId,
-      outcome: approval.kind === "deny" ? "failure" : "success",
-      durationMs: clock() - startMs,
-      metadata: meta as JsonObject,
-    };
-    try {
-      approvalSink(ctx.session.sessionId as string, step);
-    } catch (e: unknown) {
-      swallowError(e, { package: PKG, operation: "approval-step" });
-    }
-  }
+  // Wire up filterTools via factory
+  const { filterTools } = createFilterTools({
+    config,
+    auditSink,
+    clock,
+    filterCapRecordedKeys,
+    getTracker,
+    getTurnSoftDenyCounter,
+    getSoftDenyLog,
+    isDefaultDenyLike,
+    queryForTool,
+    resolveBatch,
+  });
 
-  /** Audit a permission decision at model-time filtering (wrapModelCall). */
-  function auditFilterDecision(
-    ctx: TurnContext,
-    resource: string,
-    decision: PermissionDecision,
-    sink: AuditSink,
-  ): void {
-    const entry: AuditEntry = {
-      schema_version: 2,
-      timestamp: clock(),
-      sessionId: ctx.session.sessionId as string,
-      agentId: ctx.session.agentId,
-      turnIndex: ctx.turnIndex,
-      kind: "permission_decision",
-      durationMs: 0,
-      metadata: {
-        permissionCheck: true,
-        phase: "filter",
-        resource,
-        effect: decision.effect,
-        ...(decision.effect !== "allow" ? { reason: decision.reason } : {}),
-      },
-    };
-    void sink.log(entry).catch((e: unknown) => {
-      swallowError(e, { package: PKG, operation: "audit" });
-    });
-  }
+  // Wire up handleAskDecision via factory
+  const { handleAskDecision } = createHandleAskDecision({
+    config,
+    auditSink,
+    clock,
+    approvalTimeoutMs,
+    persistentStore,
+    persistentAgentId,
+    backendFingerprint,
+    inflightApprovals,
+    inflightKeysBySession,
+    alwaysAllowedBySession,
+    getTracker,
+    getApprovalCache,
+    emitApprovalStep,
+    auditApprovalOutcome,
+    computeApprovalCacheKey,
+    serializeTurnContext,
+  });
 
-  async function resolveDecision(
-    query: PermissionQuery,
-    sessionId: string,
-  ): Promise<PermissionDecision> {
-    // Denial escalation: skip backend if this tool+context has enough recent policy denials.
-    // If the query context is not serializable (cacheKey undefined), skip escalation
-    // entirely — we cannot scope it safely and must not match across contexts.
-    if (escalationEnabled) {
-      const cacheKey = decisionCacheKey(query);
-      if (cacheKey !== undefined) {
-        const tracker = getTracker(sessionId);
-        const now = clock();
-        const cutoff = escalationWindowMs > 0 ? now - escalationWindowMs : 0;
-        const recentPolicyDenials = tracker
-          .getByTool(query.resource)
-          .filter(
-            (r) =>
-              r.source === "policy" &&
-              r.softness !== "soft" &&
-              r.origin !== "soft-conversion" &&
-              r.timestamp >= cutoff &&
-              (r.queryKey === undefined || r.queryKey === cacheKey),
-          );
-        if (recentPolicyDenials.length >= escalationThreshold) {
-          return {
-            effect: "deny",
-            reason: `Auto-denied: ${escalationThreshold}+ prior denials this session`,
-            [IS_ESCALATED]: true,
-          } as PermissionDecision;
-        }
-      }
-    }
-
-    // Circuit breaker check
-    if (cb !== undefined && !cb.isAllowed()) {
-      return failClosedDeny("Permission backend circuit open — failing closed");
-    }
-
-    // Cache check (per-session, skip if key not serializable)
-    const decisionCache = getDecisionCache(sessionId);
-    const cacheKey = decisionCacheKey(query);
-    if (decisionCache !== undefined && cacheKey !== undefined) {
-      const cached = decisionCache.get(cacheKey);
-      if (cached !== undefined) return tagCached(cached);
-    }
-
-    try {
-      const raw = await backend.check(query);
-      const decision = validateDecision(raw);
-
-      // Malformed response = backend failure for circuit breaker purposes
-      if (cb !== undefined) {
-        if (decision === FAIL_CLOSED_DENY) {
-          cb.recordFailure();
-        } else {
-          cb.recordSuccess();
-        }
-      }
-      if (decisionCache !== undefined && decision !== FAIL_CLOSED_DENY && cacheKey !== undefined) {
-        decisionCache.set(cacheKey, decision);
-      }
-      return decision;
-    } catch (e: unknown) {
-      if (cb !== undefined) cb.recordFailure();
-      return failClosedDeny(
-        `Permission backend error — failing closed: ${e instanceof Error ? e.message : String(e)}`,
-      );
-    }
-  }
-
-  async function resolveBatch(
-    queries: readonly PermissionQuery[],
-    sessionId: string,
-  ): Promise<readonly PermissionDecision[]> {
-    if (queries.length === 0) return [];
-
-    const decisionCache = getDecisionCache(sessionId);
-
-    // Partition into cached, escalated, and uncached
-    const results: (PermissionDecision | undefined)[] = new Array(queries.length).fill(undefined);
-    const uncachedIndices: number[] = [];
-
-    // Denial escalation: resolve escalated tools before cache/backend.
-    // Skip escalation for queries with non-serializable context (undefined cacheKey).
-    if (escalationEnabled) {
-      const tracker = getTracker(sessionId);
-      const now = clock();
-      const cutoff = escalationWindowMs > 0 ? now - escalationWindowMs : 0;
-      for (let i = 0; i < queries.length; i++) {
-        const query = queries[i];
-        if (query === undefined) continue;
-        const cacheKey = decisionCacheKey(query);
-        if (cacheKey === undefined) continue;
-        const recentPolicyDenials = tracker
-          .getByTool(query.resource)
-          .filter(
-            (r) =>
-              r.source === "policy" &&
-              r.softness !== "soft" &&
-              r.origin !== "soft-conversion" &&
-              r.timestamp >= cutoff &&
-              (r.queryKey === undefined || r.queryKey === cacheKey),
-          );
-        if (recentPolicyDenials.length >= escalationThreshold) {
-          results[i] = {
-            effect: "deny",
-            reason: `Auto-denied: ${escalationThreshold}+ prior denials this session`,
-            [IS_ESCALATED]: true,
-          } as PermissionDecision;
-        }
-      }
-    }
-
-    if (decisionCache !== undefined) {
-      for (let i = 0; i < queries.length; i++) {
-        if (results[i] !== undefined) continue; // already escalated
-        const query = queries[i];
-        if (query === undefined) continue;
-        const key = decisionCacheKey(query);
-        const cached = key !== undefined ? decisionCache.get(key) : undefined;
-        if (cached !== undefined) {
-          results[i] = tagCached(cached);
-        } else {
-          uncachedIndices.push(i);
-        }
-      }
-    } else {
-      for (let i = 0; i < queries.length; i++) {
-        if (results[i] !== undefined) continue; // already escalated
-        uncachedIndices.push(i);
-      }
-    }
-
-    if (uncachedIndices.length === 0) {
-      return results as readonly PermissionDecision[];
-    }
-
-    // Circuit breaker
-    if (cb !== undefined && !cb.isAllowed()) {
-      const deny = failClosedDeny("Permission backend circuit open — failing closed");
-      for (const i of uncachedIndices) {
-        results[i] = deny;
-      }
-      return results as readonly PermissionDecision[];
-    }
-
-    // Resolve uncached queries
-    // biome-ignore lint/style/noNonNullAssertion: uncachedIndices only contains valid queries indices
-    const uncachedQueries = uncachedIndices.map((i) => queries[i]!);
-    try {
-      let rawDecisions: readonly unknown[];
-      if (backend.checkBatch !== undefined) {
-        rawDecisions = (await backend.checkBatch(uncachedQueries)) as readonly unknown[];
-      } else {
-        rawDecisions = (await Promise.all(
-          uncachedQueries.map((q) => backend.check(q)),
-        )) as readonly unknown[];
-      }
-
-      // Validate batch length — fail closed on mismatch (counts as backend failure)
-      if (!Array.isArray(rawDecisions) || rawDecisions.length !== uncachedQueries.length) {
-        if (cb !== undefined) cb.recordFailure();
-        for (const i of uncachedIndices) {
-          results[i] = FAIL_CLOSED_DENY;
-        }
-        return results as readonly PermissionDecision[];
-      }
-
-      // Validate all decisions first — any malformed element poisons the
-      // entire batch (fail-closed: a corrupted backend response cannot be
-      // partially trusted at a permission boundary)
-      const validated: PermissionDecision[] = [];
-      let hasValidationFailure = false;
-      for (let j = 0; j < uncachedIndices.length; j++) {
-        const decision = validateDecision(rawDecisions[j]);
-        if (decision === FAIL_CLOSED_DENY) hasValidationFailure = true;
-        validated.push(decision);
-      }
-
-      if (hasValidationFailure) {
-        // Poison entire batch — deny all uncached queries, cache nothing
-        if (cb !== undefined) cb.recordFailure();
-        for (const i of uncachedIndices) {
-          results[i] = FAIL_CLOSED_DENY;
-        }
-      } else {
-        if (cb !== undefined) cb.recordSuccess();
-        for (let j = 0; j < uncachedIndices.length; j++) {
-          // biome-ignore lint/style/noNonNullAssertion: j < uncachedIndices.length && validated.length === uncachedIndices.length
-          const idx = uncachedIndices[j]!;
-          // biome-ignore lint/style/noNonNullAssertion: j < validated.length (validated built from same loop)
-          const decision = validated[j]!;
-          results[idx] = decision;
-          if (decisionCache !== undefined) {
-            // biome-ignore lint/style/noNonNullAssertion: idx is a valid index from uncachedIndices
-            const key = decisionCacheKey(queries[idx]!);
-            if (key !== undefined) decisionCache.set(key, decision);
-          }
-        }
-      }
-    } catch (e: unknown) {
-      if (cb !== undefined) cb.recordFailure();
-      const deny = failClosedDeny(
-        `Permission backend error — failing closed: ${e instanceof Error ? e.message : String(e)}`,
-      );
-      for (const i of uncachedIndices) {
-        results[i] = deny;
-      }
-    }
-
-    return results as readonly PermissionDecision[];
-  }
-
-  async function filterTools(ctx: TurnContext, request: ModelRequest): Promise<ModelRequest> {
-    const tools = request.tools;
-    if (tools === undefined || tools.length === 0) return request;
-
-    // Include model request metadata so filtering uses the same policy
-    // inputs as execution-time wrapToolCall (prevents visibility/auth mismatch)
-    const queries = tools.map((t) => queryForTool(ctx, t.name, request.metadata));
-    const decisions = await resolveBatch(queries, ctx.session.sessionId as string);
-
-    const sessionTracker = getTracker(ctx.session.sessionId as string);
-
-    // #1650 loop round-9: capture the FINAL enforced decision per tool so
-    // reportDecision's filteredTools summary reflects what was actually
-    // emitted (hardened reason for soft→hard conversions), not the
-    // pre-conversion soft decision.
-    const enforcedDecisionByIndex = new Map<number, PermissionDecision>();
-
-    // Tools whose per-command policy is enforced at wrapToolCall via
-    // bash prefix enrichment — keep visible at model-time so the
-    // execution-time prefix rules are reachable UNLESS the operator
-    // has explicitly denied the plain tool id. An explicit deny
-    // (non-default-deny) on `bash` must still filter the tool out so
-    // operators can turn off bash wholesale via policy.
-    const bashVisibleSet = new Set(config.bashVisibleTools ?? []);
-    const bypassFilter = (toolName: string, decision: PermissionDecision): boolean => {
-      if (config.resolveBashCommand === undefined) return false;
-      if (!bashVisibleSet.has(toolName)) return false;
-      // Explicit deny overrides the visibility bypass. Default-deny
-      // (no rule matches) does NOT — that's exactly the case
-      // bashVisibleTools exists to accommodate.
-      if (decision.effect === "deny" && !isDefaultDenyLike(decision)) return false;
-      return true;
-    };
-
-    const filtered = tools.filter((tool, i) => {
-      // biome-ignore lint/style/noNonNullAssertion: decisions.length === tools.length (resolveBatch returns same length)
-      const decision = decisions[i]!;
-      // biome-ignore lint/style/noNonNullAssertion: queries built from tools.map — same length as filter callback index
-      const query = queries[i]!;
-      const sid = ctx.session.sessionId as string;
-
-      if (bypassFilter(tool.name, decision)) {
-        // Pass through. Record an allow decision for observability so
-        // reporting/audit still see the tool was offered to the model.
-        enforcedDecisionByIndex.set(i, { effect: "allow" });
-        return true;
-      }
-
-      // #1650 loop round-4: audit/dispatch fire AFTER the hard-conversion
-      // checks so observers see the FINAL decision shape, not the original
-      // soft decision that was about to be hard-converted. Keeps audit sinks
-      // consistent with what was actually enforced.
-      const emit = (finalDecision: PermissionDecision): void => {
-        enforcedDecisionByIndex.set(i, finalDecision);
-        if (auditSink !== undefined) {
-          auditFilterDecision(ctx, tool.name, finalDecision, auditSink);
-        }
-        void ctx.dispatchPermissionDecision?.(query, finalDecision);
-      };
-
-      if (decision.effect === "deny") {
-        const dispositionIsSoft = (decision.disposition ?? "hard") === "soft";
-        if (dispositionIsSoft) {
-          const cacheKey = decisionCacheKey(query);
-          // #1650 loop round-3/4: unkeyable context → hard-convert (mirror
-          // execute-time fail-closed). Record in DenialTracker with
-          // origin: "soft-conversion" (same vocabulary as execute-time) so
-          // observers see a consistent hard denial trail.
-          if (cacheKey === undefined) {
-            const hardened: PermissionDecision = {
-              ...decision,
-              disposition: "hard",
-              reason: `${decision.reason} (unkeyable context — failing closed)`,
-            };
-            // #1650 loop round-9: dedup unkeyable filter-time records per
-            // (session, turn, toolId). Repeated planning passes in the same
-            // turn with unkeyable context would otherwise churn the 1024-entry
-            // DenialTracker FIFO and evict native hard-deny history.
-            const unkeyableRecordKey = `${sid}\0${ctx.turnIndex}\0unkeyable\0${tool.name}`;
-            if (!filterCapRecordedKeys.has(unkeyableRecordKey)) {
-              filterCapRecordedKeys.add(unkeyableRecordKey);
-              sessionTracker.record({
-                toolId: tool.name,
-                reason: decision.reason,
-                timestamp: clock(),
-                principal: ctx.session.agentId,
-                turnIndex: ctx.turnIndex,
-                source: denialSource(decision),
-                queryKey: undefined,
-                softness: "hard",
-                origin: "soft-conversion",
-              });
-            }
-            emit(hardened);
-            return false;
-          }
-          // #1650 loop round-2/6: if the soft-deny counter is ALREADY at or
-          // over cap for this tool in the current turn, stop advertising the
-          // tool so the model doesn't burn iterations on guaranteed-hard
-          // failures.
-          //
-          // Trade-off (loop rounds 6→9): previously used a tool-identity
-          // prefix peek to catch path-sensitive cap exhaustion, but that
-          // over-stripped tools for unrelated intents (different paths) whose
-          // future call would have been allowed. Reverting to exact-key peek
-          // — path-sensitive cap exhaustion is invisible at planning time;
-          // execute-time still enforces (hard-throws on the Nth+1 exact-key
-          // deny) and the outer engine max-iterations bounds the worst case.
-          // For tools WITHOUT a resolveToolPath callback, filter-time keys
-          // match execute-time exactly and this check strips correctly.
-          const cap = config.softDenyPerTurnCap ?? DEFAULT_SOFT_DENY_PER_TURN_CAP;
-          const turnScopedKey = `${ctx.turnIndex}\0${cacheKey}`;
-          const currentCount = getTurnSoftDenyCounter(sid).peek(turnScopedKey);
-          if (currentCount >= cap) {
-            // #1650 loop round-3: record DenialTracker ONCE per (session, turn,
-            // cacheKey). Repeated planning passes in the same turn would
-            // otherwise evict native hard-deny history from the bounded FIFO.
-            const capRecordKey = `${sid}\0${ctx.turnIndex}\0${cacheKey}`;
-            if (!filterCapRecordedKeys.has(capRecordKey)) {
-              filterCapRecordedKeys.add(capRecordKey);
-              sessionTracker.record({
-                toolId: tool.name,
-                reason: decision.reason,
-                timestamp: clock(),
-                principal: ctx.session.agentId,
-                turnIndex: ctx.turnIndex,
-                source: denialSource(decision),
-                queryKey: cacheKey,
-                softness: "hard",
-                origin: "soft-conversion",
-              });
-            }
-            // Always emit the hardened decision so audit/observers see the
-            // final enforced shape (even if tracker dedup suppressed the
-            // record-write above).
-            const hardened: PermissionDecision = {
-              ...decision,
-              disposition: "hard",
-              reason: `${decision.reason} (soft-deny retry cap ${cap} exceeded this turn)`,
-            };
-            emit(hardened);
-            return false;
-          }
-          // Soft-deny → keep tool visible. Record in isolated SoftDenyLog so
-          // high-volume soft denies don't evict native hard-deny history from
-          // the shared DenialTracker budget. Emit the ORIGINAL soft decision
-          // (this is what actually happens at planning time — the tool stays
-          // visible).
-          getSoftDenyLog(sid).record({
-            toolId: tool.name,
-            reason: decision.reason,
-            timestamp: clock(),
-            principal: ctx.session.agentId,
-            turnIndex: ctx.turnIndex,
-            queryKey: cacheKey,
-          });
-          emit(decision);
-          return true;
-        }
-        // Hard-deny: existing behavior — record with softness+origin, emit, strip.
-        sessionTracker.record({
-          toolId: tool.name,
-          reason: decision.reason,
-          timestamp: clock(),
-          principal: ctx.session.agentId,
-          turnIndex: ctx.turnIndex,
-          source: denialSource(decision),
-          queryKey: decisionCacheKey(query),
-          softness: "hard",
-          origin: "native",
-        });
-        emit(decision);
-        return false;
-      }
-      // allow/ask — emit the original decision so audit/observer chain sees it.
-      emit(decision);
-      return true;
-    });
-
-    const filteredCount = tools.length - filtered.length;
-    if (filteredCount > 0) {
-      // #1650 loop round-9: prefer the enforced (possibly hardened) decision
-      // captured during emit over the raw backend decision, so filteredTools
-      // reflects the final reason ("unkeyable context — failing closed" /
-      // "soft-deny retry cap N exceeded this turn") rather than the original
-      // soft-policy reason.
-      const filteredDetails = tools
-        .map((t, i) => {
-          const decision = enforcedDecisionByIndex.get(i) ?? decisions[i];
-          return { name: t.name, decision };
-        })
-        .filter(
-          (
-            d,
-          ): d is {
-            readonly name: string;
-            readonly decision: { readonly effect: "deny"; readonly reason: string };
-          } => d.decision?.effect === "deny",
-        )
-        .map((d) => ({
-          tool: d.name,
-          reason: d.decision.reason,
-          source: denialSource(d.decision),
-        }));
-      ctx.reportDecision?.({
-        phase: "filter",
-        totalTools: tools.length,
-        allowedCount: filtered.length,
-        filteredCount,
-        filteredTools: filteredDetails,
-      });
-    } else {
-      ctx.reportDecision?.({
-        phase: "filter",
-        totalTools: tools.length,
-        allowedCount: tools.length,
-        filteredCount: 0,
-      });
-    }
-    if (filtered.length === tools.length) return request;
-    return { ...request, tools: filtered };
-  }
+  // Wire up wrapToolCall via factory
+  const { wrapToolCall } = createWrapToolCall({
+    config,
+    auditSink,
+    clock,
+    backendSupportsDualKey,
+    enrichResource,
+    queryForTool,
+    resolveDecision,
+    strictestDecision,
+    auditDecision,
+    getTracker,
+    getTurnSoftDenyCounter,
+    getSoftDenyLog,
+    handleAskDecision,
+  });
 
   // -----------------------------------------------------------------------
   // Middleware + Handle
@@ -1513,310 +685,7 @@ export function createPermissionsMiddleware(
       request: ToolRequest,
       next: ToolHandler,
     ): Promise<ToolResponse> {
-      // `callId` is a UI/observability identifier carried on a dedicated
-      // ToolRequest.callId field (NOT inside `metadata`), so it never
-      // enters the backend policy query, the approval cache key, or
-      // the in-flight dedup key. (#1759 review round 6)
-      // Resolve file path for fs tools so permission rules can match on context.path.
-      const resolvedPath = config.resolveToolPath?.(request.toolId, request.input);
-      const { policy: enrichedResource, grant: grantKey } = enrichResource(
-        request.toolId,
-        request.input,
-      );
-      const enrichedQuery = queryForTool(ctx, enrichedResource, request.metadata, resolvedPath);
-      const startMs = clock();
-      // Dual-key evaluation: when the enriched resource differs from
-      // the plain tool id, consult the backend for BOTH and take the
-      // stricter decision. The "effective" resource/query carries
-      // forward whichever side won so denial tracking, soft-deny
-      // caps, and audit records aggregate under the right bucket —
-      // a plain `deny: bash` must aggregate across all subcommand
-      // variants, not fragment into per-prefix buckets.
-      // `resource` is the observable identity — used for audit,
-      // reportDecision, and approval-step payloads. It stays on the
-      // enriched form so per-command governance is visible even
-      // when a legacy plain-tool rule ultimately produced the
-      // decision.
-      //
-      // `trackingResource` is the denial-tracker + soft-deny bucket.
-      // When a plain-tool deny wins, it aggregates subcommand
-      // variants under the plain id so retry caps and escalation
-      // fire consistently. When the enriched decision wins it uses
-      // the enriched resource for per-prefix fidelity.
-      // Dual-key merge (marker-aware backends): issue both enriched and
-      // plain queries, merge via `strictestDecision`. Unmarked
-      // fall-through denies on one key yield to an explicit allow/ask
-      // on the other, preserving legacy plain-tool rules while
-      // honoring per-prefix policy.
-      //
-      // Single-key fallback (legacy backends without the marker flag):
-      // use only the plain tool id. Safe — avoids hard-denying via
-      // unmarked fall-through. Construction emitted a warning so
-      // operators know enrichment isn't enforced.
-      const resource = enrichedResource;
-      let trackingResource = enrichedResource;
-      let query = enrichedQuery;
-      let decision: PermissionDecision;
-      if (!backendSupportsDualKey && enrichedResource !== request.toolId) {
-        const plainQuery = queryForTool(ctx, request.toolId, request.metadata, resolvedPath);
-        query = plainQuery;
-        trackingResource = request.toolId;
-        decision = await resolveDecision(query, ctx.session.sessionId as string);
-      } else {
-        decision = await resolveDecision(query, ctx.session.sessionId as string);
-        if (enrichedResource !== request.toolId) {
-          const plainQuery = queryForTool(ctx, request.toolId, request.metadata, resolvedPath);
-          const plain = await resolveDecision(plainQuery, ctx.session.sessionId as string);
-          const combined = strictestDecision(plain, decision);
-          if (combined === plain) {
-            trackingResource = request.toolId;
-            query = plainQuery;
-          }
-          decision = combined;
-        }
-      }
-      // Structural-complexity ratchet: `!complex` covers compound
-      // commands, redirections, subshells, command substitution, env
-      // -S, and anything canonicalPrefix could not safely reduce. A
-      // broad `allow: bash:*` must NOT silently authorize these —
-      // they can hide side effects or nested dangerous payloads.
-      //
-      // But operators who INTENTIONALLY whitelist compound forms
-      // (`allow: ["bash:!complex*"]`) should be honored. Disambiguate
-      // by probing with a nonsense resource that a wildcard would
-      // still match but an explicit `bash:!complex*` rule would not.
-      // If the probe also allows, the original allow came from a
-      // wildcard — ratchet to ask. If the probe denies, the operator
-      // has an explicit `!complex` rule — honor it.
-      if (
-        decision.effect === "allow" &&
-        enrichedResource !== request.toolId &&
-        enrichedResource === `${request.toolId}:${UNSAFE_PREFIX}`
-      ) {
-        // Probe under the PLAIN tool id so ONLY truly broad
-        // wildcards like `bash:*` match it. Narrower rules
-        // (`bash:!complex*`, `bash:sudo*`, `bash:python*`) do not
-        // match this probe shape, so operators who explicitly
-        // whitelist specific prefixes are honored.
-        const probeQuery = queryForTool(
-          ctx,
-          `${request.toolId}:__broad_wildcard_probe__`,
-          request.metadata,
-          resolvedPath,
-        );
-        const probe = await resolveDecision(probeQuery, ctx.session.sessionId as string);
-        const fromWildcard = probe.effect === "allow";
-        if (fromWildcard) {
-          decision = {
-            effect: "ask",
-            reason: "complex/unparseable shell form requires review",
-          };
-        }
-      }
-      // Dangerous-command ratchet: if the raw command matches ANY
-      // DANGEROUS_PATTERN and the combined decision is "allow",
-      // upgrade to "ask" so a human reviews it. Broad allow rules
-      // like `allow: bash:*` cannot silently authorize
-      // structural-danger forms.
-      //
-      // BUT: operators who explicitly allow a dangerous prefix
-      // (e.g. `allow: bash:sudo`, `allow: bash:python*`,
-      // `allow: bash:chmod*`) have opted in deliberately — typically
-      // for headless/non-interactive automation. Probe under the
-      // plain tool id so ONLY truly broad wildcards match the probe.
-      if (decision.effect === "allow" && config.resolveBashCommand !== undefined) {
-        const raw = config.resolveBashCommand(request.toolId, request.input);
-        if (raw !== undefined && raw.trim().length > 0) {
-          const danger = classifyCommand(raw.trim());
-          if (danger.severity !== null) {
-            const probeQuery = queryForTool(
-              ctx,
-              `${request.toolId}:__broad_wildcard_probe__`,
-              request.metadata,
-              resolvedPath,
-            );
-            const probe = await resolveDecision(probeQuery, ctx.session.sessionId as string);
-            const fromWildcard = probe.effect === "allow";
-            if (fromWildcard) {
-              const categories = Array.from(new Set(danger.matchedPatterns.map((p) => p.category)));
-              decision = {
-                effect: "ask",
-                reason: `dangerous command pattern matched (${categories.join(", ")})`,
-              };
-            }
-          }
-        }
-      }
-      const durationMs = clock() - startMs;
-
-      // Non-deny paths: audit + dispatch here. Deny paths handle their own
-      // audit, dispatch, and report inside the deny branch (Task 10 of #1650)
-      // so that soft vs hard-converted decisions use the correct final decision object.
-      if (decision.effect !== "deny") {
-        if (auditSink !== undefined) {
-          auditDecision(ctx, resource, decision, durationMs, auditSink);
-        }
-        // Allow/ask: fire-and-forget dispatch here.
-        void ctx.dispatchPermissionDecision?.(query, decision);
-        ctx.reportDecision?.({
-          phase: "execute",
-          toolId: request.toolId,
-          resource,
-          toolInput: safePreviewJson(request.input, 300),
-          action: decision.effect,
-          durationMs,
-          ...(decision.effect !== "allow" ? { reason: decision.reason } : {}),
-          source: denialSource(decision),
-        });
-      }
-
-      if (decision.effect === "deny") {
-        const source = denialSource(decision);
-        const disposition = decision.disposition ?? "hard";
-
-        const isSoftCandidate =
-          disposition === "soft" &&
-          source !== "approval" &&
-          !isFailClosed(decision) &&
-          !isEscalated(decision); // IS_CACHED does NOT set IS_ESCALATED — cached replays stay soft-eligible
-
-        const sessionId = ctx.session.sessionId as string;
-
-        type DenyDecision = Extract<PermissionDecision, { readonly effect: "deny" }>;
-
-        const hardConvertedDecision = (suffix: string): DenyDecision => ({
-          ...decision,
-          disposition: "hard" as const,
-          reason: `${decision.reason} (${suffix})`,
-        });
-
-        const emitDenyAudit = (finalDecision: DenyDecision): void => {
-          if (auditSink !== undefined) {
-            auditDecision(ctx, resource, finalDecision, durationMs, auditSink);
-          }
-          void ctx.dispatchPermissionDecision?.(query, finalDecision);
-          ctx.reportDecision?.({
-            phase: "execute",
-            toolId: request.toolId,
-            resource,
-            toolInput: safePreviewJson(request.input, 300),
-            action: "deny",
-            durationMs,
-            reason: finalDecision.reason,
-            source: denialSource(finalDecision),
-          });
-        };
-
-        if (isSoftCandidate) {
-          const cacheKey = decisionCacheKey(query);
-
-          if (cacheKey === undefined) {
-            // Unkeyable context — cannot scope the soft-deny counter safely → fail closed.
-            const hardened = hardConvertedDecision("unkeyable context — failing closed");
-            getTracker(sessionId).record({
-              toolId: trackingResource,
-              reason: decision.reason,
-              timestamp: clock(),
-              principal: ctx.session.agentId,
-              turnIndex: ctx.turnIndex,
-              source,
-              queryKey: undefined,
-              softness: "hard",
-              origin: "soft-conversion",
-            });
-            emitDenyAudit(hardened);
-            throw new KoiRuntimeError({
-              code: "PERMISSION",
-              message: hardened.reason,
-              retryable: false,
-            });
-          }
-
-          // Per-turn cap check. Prefix the counter key with turnIndex so
-          // overlapping turns in the same session cannot reset or share each
-          // other's budget. Loop round-5 fix.
-          const cap = config.softDenyPerTurnCap ?? DEFAULT_SOFT_DENY_PER_TURN_CAP;
-          const counter = getTurnSoftDenyCounter(sessionId);
-          const turnScopedKey = `${ctx.turnIndex}\0${cacheKey}`;
-          if (counter.countAndCap(turnScopedKey, cap) === "over_cap") {
-            const hardened = hardConvertedDecision(`soft-deny retry cap ${cap} exceeded this turn`);
-            getTracker(sessionId).record({
-              toolId: trackingResource,
-              reason: decision.reason,
-              timestamp: clock(),
-              principal: ctx.session.agentId,
-              turnIndex: ctx.turnIndex,
-              source,
-              queryKey: cacheKey,
-              softness: "hard",
-              origin: "soft-conversion",
-            });
-            emitDenyAudit(hardened);
-            throw new KoiRuntimeError({
-              code: "PERMISSION",
-              message: hardened.reason,
-              retryable: false,
-            });
-          }
-
-          // Soft path: record in isolated SoftDenyLog (NOT DenialTracker).
-          getSoftDenyLog(sessionId).record({
-            toolId: trackingResource,
-            reason: decision.reason,
-            timestamp: clock(),
-            principal: ctx.session.agentId,
-            turnIndex: ctx.turnIndex,
-            queryKey: cacheKey,
-          });
-          emitDenyAudit(decision);
-          // Trust-boundary: output contains only toolId, never decision.reason.
-          // `blockedByHook: true` is the canonical downstream marker honored
-          // by event-trace, middleware-report, and session-transcript to
-          // classify this response as a non-execution rather than a successful
-          // tool call. Loop round-10 fix.
-          // `permissionDenied: true` is the specific signal for #1650 soft-deny.
-          return {
-            output: `Permission denied for tool "${request.toolId}". This tool is not available in the current scope.`,
-            metadata: {
-              isError: true,
-              blockedByHook: true,
-              permissionDenied: true,
-              hookName: "permissions",
-              toolId: request.toolId,
-            },
-          };
-        }
-
-        // Native hard path: record with origin: "native", dispatch original decision, throw.
-        getTracker(sessionId).record({
-          toolId: trackingResource,
-          reason: decision.reason,
-          timestamp: clock(),
-          principal: ctx.session.agentId,
-          turnIndex: ctx.turnIndex,
-          source,
-          queryKey: decisionCacheKey(query),
-          softness: "hard",
-          origin: "native",
-        });
-        emitDenyAudit(decision);
-        throw new KoiRuntimeError({
-          code: "PERMISSION",
-          message: decision.reason,
-          retryable: false,
-        });
-      }
-
-      if (decision.effect === "ask") {
-        // Pass a dispatch callback so each approval path fires the outcome
-        // BEFORE calling next(request) — ensures recording even if the tool throws.
-        return handleAskDecision(ctx, request, resource, grantKey, next, decision, (d) => {
-          void ctx.dispatchPermissionDecision?.(query, d);
-        });
-      }
-
-      // allow
-      return next(request);
+      return wrapToolCall(ctx, request, next);
     },
 
     async onBeforeTurn(_ctx: TurnContext): Promise<void> {
@@ -1883,516 +752,4 @@ export function createPermissionsMiddleware(
       return getTracker(sessionId);
     },
   }) as PermissionsMiddlewareHandle;
-
-  // -----------------------------------------------------------------------
-  // Ask / approval flow
-  // -----------------------------------------------------------------------
-
-  async function handleAskDecision(
-    ctx: TurnContext,
-    request: ToolRequest,
-    resource: string,
-    grantKey: string,
-    next: ToolHandler,
-    decision: PermissionDecision & { readonly effect: "ask" },
-    dispatchApprovalOutcome?: (d: PermissionDecision) => void,
-  ): Promise<ToolResponse> {
-    const approvalHandler: ApprovalHandler | undefined = ctx.requestApproval;
-
-    if (approvalHandler === undefined) {
-      throw new KoiRuntimeError({
-        code: "PERMISSION",
-        message: `Tool "${request.toolId}" requires approval but no approval handler is configured`,
-        retryable: false,
-      });
-    }
-
-    // Check persistent always-allow grants (cross-session, SQLite-backed).
-    // Fail-open: if the store throws (corrupt DB, lock contention), fall through
-    // to the session check and ultimately to the user prompt. This is fail-safe —
-    // a broken store means more prompts, not silent denials or silent allows.
-    // Persistent grants require a real user identity — anonymous sessions
-    // must not share a durable principal, so we skip the store entirely.
-    // Use persistentAgentId if configured (stable across restarts) — falls back
-    // to the per-process agentId for multi-agent runtimes.
-    const persistentUserId = ctx.session.userId;
-    const persistentAid = persistentAgentId ?? ctx.session.agentId;
-    if (persistentStore !== undefined && persistentUserId !== undefined) {
-      try {
-        // Key persistent grants by `grantKey` (exact-command hash) so
-        // approving `bash:git status` does not auto-approve the stricter
-        // `bash:git status --short` or the materially different
-        // `bash:git push --force`. Distinct argv → distinct hash → fresh
-        // prompt.
-        //
-        // Legacy fallback is OFF by default: a blanket `bash` grant
-        // stored before prefix enrichment must not retroactively
-        // authorize newly-scoped commands. Operators migrating from
-        // pre-enrichment deployments can opt in via
-        // `legacyBashGrantFallback: true` to preserve existing grants
-        // during a rollout window. Even with the flag, `!complex`
-        // structural forms and dangerous-pattern matches always force
-        // a fresh approval.
-        const hasExact = persistentStore.has(persistentUserId, persistentAid, grantKey);
-        let hasLegacy = false;
-        if (
-          !hasExact &&
-          config.legacyBashGrantFallback === true &&
-          grantKey !== request.toolId &&
-          persistentStore.has(persistentUserId, persistentAid, request.toolId)
-        ) {
-          const rawForLegacy = config.resolveBashCommand?.(request.toolId, request.input) ?? "";
-          const isComplexForm = resource === `${request.toolId}:${UNSAFE_PREFIX}`;
-          const isDangerous =
-            rawForLegacy.trim().length > 0 &&
-            classifyCommand(rawForLegacy.trim()).severity !== null;
-          hasLegacy = !isComplexForm && !isDangerous;
-        }
-        if (hasExact || hasLegacy) {
-          const persistentStartMs = clock();
-          getTracker(ctx.session.sessionId as string).record({
-            toolId: resource,
-            reason: `auto-approved (persistent always-allow grant, agent: ${ctx.session.agentId})`,
-            timestamp: persistentStartMs,
-            principal: ctx.session.agentId,
-            turnIndex: ctx.turnIndex,
-            source: "approval",
-          });
-          emitApprovalStep(
-            ctx,
-            request.toolId,
-            { kind: "always-allow", scope: "always" },
-            request.input,
-            persistentStartMs,
-          );
-          if (auditSink !== undefined) {
-            auditApprovalOutcome(
-              ctx,
-              resource,
-              { kind: "always-allow", scope: "always" },
-              request.input,
-              clock() - persistentStartMs,
-              auditSink,
-              /* coalesced */ false,
-              /* remembered */ true,
-            );
-          }
-          // Dispatch before next() so the permission outcome is recorded even if the tool throws
-          dispatchApprovalOutcome?.({ effect: "allow" });
-          return next(request);
-        }
-      } catch {
-        // Fall through to session/cache/prompt — fail-open.
-      }
-    }
-
-    // Check always-allowed set (from prior "always-allow" decisions).
-    // Session bypass is keyed by agentId + grantKey (exact-command
-    // hash). A user's "a" press approves THIS command + argv, not every
-    // later invocation of the same prefix. When enrichment is off,
-    // grantKey falls back to the plain tool id, preserving the original
-    // one-approve-per-tool behavior.
-    const alwaysAllowKey = `${ctx.session.agentId}\0${grantKey}`;
-    const sessionAlwaysAllowed = alwaysAllowedBySession.get(ctx.session.sessionId as string);
-    if (sessionAlwaysAllowed?.has(alwaysAllowKey)) {
-      const alwaysAllowStartMs = clock();
-      getTracker(ctx.session.sessionId as string).record({
-        toolId: resource,
-        reason: `auto-approved (always-allow session rule, agent: ${ctx.session.agentId})`,
-        timestamp: alwaysAllowStartMs,
-        principal: ctx.session.agentId,
-        turnIndex: ctx.turnIndex,
-        source: "approval",
-      });
-      emitApprovalStep(
-        ctx,
-        request.toolId,
-        { kind: "always-allow", scope: "session" },
-        request.input,
-        alwaysAllowStartMs,
-      );
-      // Dispatch before next() so the permission outcome is recorded even if the tool throws
-      dispatchApprovalOutcome?.({ effect: "allow" });
-      return next(request);
-    }
-
-    // Check approval cache (per-session)
-    const approvalCache = getApprovalCache(ctx.session.sessionId as string);
-    if (approvalCache !== undefined) {
-      const userId = ctx.session.userId ?? "__anonymous__";
-      const ctxStr = serializeTurnContext(ctx);
-      const cacheKey = computeApprovalCacheKey(
-        backendFingerprint,
-        ctx.session.sessionId as string,
-        userId,
-        ctx.session.agentId,
-        request.toolId,
-        request.input,
-        ctxStr,
-        request.metadata,
-        decision.reason,
-        grantKey,
-      );
-
-      if (cacheKey !== undefined && approvalCache.has(cacheKey)) {
-        emitApprovalStep(ctx, request.toolId, { kind: "allow" }, request.input, clock());
-        // Dispatch before next() so the permission outcome is recorded even if the tool throws
-        dispatchApprovalOutcome?.({ effect: "allow" });
-        return next(request);
-      }
-    }
-
-    // Build dedup key for in-flight coordination
-    const dedupUserId = ctx.session.userId ?? "__anonymous__";
-    const dedupCtx = serializeTurnContext(ctx);
-    const dedupKey = computeApprovalCacheKey(
-      backendFingerprint,
-      ctx.session.sessionId as string,
-      dedupUserId,
-      ctx.session.agentId,
-      request.toolId,
-      request.input,
-      dedupCtx,
-      request.metadata,
-      decision.reason,
-      grantKey,
-    );
-
-    // Coalesce concurrent identical asks onto a single pending approval
-    if (dedupKey !== undefined) {
-      const inflight = inflightApprovals.get(dedupKey);
-      if (inflight !== undefined) {
-        // Another call is already waiting for approval — wait for its result
-        const coalescedStartMs = clock();
-        // let: rawResult is assigned in try, used after
-        let rawResult: unknown;
-        try {
-          rawResult = await inflight;
-        } catch (e: unknown) {
-          // Leader timed out or handler threw — emit failure step for this follower
-          const reason =
-            e instanceof KoiRuntimeError && e.code === "TIMEOUT" ? "timeout" : "handler_error";
-          emitApprovalStep(
-            ctx,
-            request.toolId,
-            { kind: "deny", reason },
-            request.input,
-            coalescedStartMs,
-            true,
-          );
-          if (e instanceof KoiRuntimeError) throw e;
-          throw new KoiRuntimeError({
-            code: "INTERNAL",
-            message: `Coalesced approval error for "${request.toolId}"`,
-            retryable: false,
-            cause: e,
-          });
-        }
-        const result = validateApprovalDecision(rawResult);
-
-        // Emit approval-outcome audit + trajectory for this coalesced caller.
-        // Marked coalesced: true so downstream systems know this reused an existing
-        // human decision rather than prompting a new one.
-        const coalescedDurationMs = clock() - coalescedStartMs;
-        if (result !== undefined && auditSink !== undefined) {
-          auditApprovalOutcome(
-            ctx,
-            resource,
-            result,
-            request.input,
-            coalescedDurationMs,
-            auditSink,
-            true,
-          );
-        }
-        if (result !== undefined) {
-          emitApprovalStep(ctx, request.toolId, result, request.input, coalescedStartMs, true);
-        } else {
-          // Malformed coalesced response — emit failure step so it's observable
-          emitApprovalStep(
-            ctx,
-            request.toolId,
-            { kind: "deny", reason: "malformed_response" },
-            request.input,
-            coalescedStartMs,
-            true,
-          );
-        }
-
-        if (result === undefined || result.kind === "deny") {
-          dispatchApprovalOutcome?.({
-            effect: "deny",
-            reason: `Tool "${request.toolId}" denied (coalesced approval)`,
-          });
-          throw new KoiRuntimeError({
-            code: "PERMISSION",
-            message: `Tool "${request.toolId}" denied (coalesced approval)`,
-            retryable: false,
-          });
-        }
-        // Dispatch allow before next() so outcome is recorded even if the tool throws
-        dispatchApprovalOutcome?.({ effect: "allow" });
-        if (result.kind === "modify") {
-          return next({ ...request, input: result.updatedInput });
-        }
-        return next(request);
-      }
-    }
-
-    // Request approval with timeout
-    const approvalStartMs = clock();
-    const ac = new AbortController();
-
-    // When approvalTimeoutMs is Infinity (default, see #1759), the timeout
-    // leg is omitted entirely — users get unbounded time to respond to
-    // interactive permission prompts. Agent-to-agent callers that need a
-    // hung-handler backstop should pass a finite value explicitly.
-    const approvalRace: readonly Promise<unknown>[] = [
-      approvalHandler({
-        toolId: request.toolId,
-        input: request.input,
-        reason: decision.reason,
-        ...(request.metadata !== undefined ? { metadata: request.metadata } : {}),
-        // Forward the per-invocation correlation id (UI-only — never
-        // touches policy or cache identity). The TUI permission bridge
-        // reads this to dispatch a per-call timer reset. (#1759 round 6)
-        ...(request.callId !== undefined ? { callId: request.callId } : {}),
-      }),
-      ...(Number.isFinite(approvalTimeoutMs)
-        ? [
-            new Promise<never>((_, reject) => {
-              const timerId = setTimeout(() => {
-                reject(
-                  new KoiRuntimeError({
-                    code: "TIMEOUT",
-                    message: `Approval for "${request.toolId}" timed out after ${approvalTimeoutMs}ms`,
-                    retryable: false,
-                  }),
-                );
-              }, approvalTimeoutMs);
-              ac.signal.addEventListener("abort", () => clearTimeout(timerId), { once: true });
-            }),
-          ]
-        : []),
-      // Race against the turn/session abort signal so an aborted turn
-      // (Ctrl+C / agent:clear) cannot win approval and execute the tool
-      // in what the user now believes is a fresh session.
-      ...(ctx.signal !== undefined
-        ? (() => {
-            // Capture signal before the Promise closure so TypeScript narrows it
-            // to AbortSignal (not AbortSignal | undefined) inside the callback.
-            const turnSignal = ctx.signal;
-            return [
-              new Promise<never>((_, reject) => {
-                if (turnSignal.aborted) {
-                  reject(
-                    new KoiRuntimeError({
-                      code: "PERMISSION",
-                      message: `Approval for "${request.toolId}" cancelled: turn was aborted`,
-                      retryable: false,
-                    }),
-                  );
-                  return;
-                }
-                turnSignal.addEventListener(
-                  "abort",
-                  () =>
-                    reject(
-                      new KoiRuntimeError({
-                        code: "PERMISSION",
-                        message: `Approval for "${request.toolId}" cancelled: turn was aborted`,
-                        retryable: false,
-                      }),
-                    ),
-                  { once: true },
-                );
-              }),
-            ];
-          })()
-        : []),
-    ];
-
-    const approvalPromise = Promise.race(approvalRace).finally(() => {
-      ac.abort();
-      if (dedupKey !== undefined) {
-        inflightApprovals.delete(dedupKey);
-        // Also remove from per-session index
-        inflightKeysBySession.get(ctx.session.sessionId as string)?.delete(dedupKey);
-      }
-    });
-
-    // Register in-flight so concurrent callers coalesce
-    if (dedupKey !== undefined) {
-      inflightApprovals.set(dedupKey, approvalPromise);
-      // Track under session so clearSessionApprovals can evict on reset
-      const sid = ctx.session.sessionId as string;
-      let keys = inflightKeysBySession.get(sid);
-      if (keys === undefined) {
-        keys = new Set();
-        inflightKeysBySession.set(sid, keys);
-      }
-      keys.add(dedupKey);
-    }
-
-    // let: tracks whether an approval step was already emitted in the try block
-    let stepEmitted = false;
-    try {
-      const rawResult = await approvalPromise;
-
-      // Validate approval response at trust boundary — fail closed on malformed
-      const approvalResult = validateApprovalDecision(rawResult);
-      if (approvalResult === undefined) {
-        emitApprovalStep(
-          ctx,
-          request.toolId,
-          { kind: "deny", reason: "malformed_response" },
-          request.input,
-          approvalStartMs,
-        );
-        stepEmitted = true;
-        dispatchApprovalOutcome?.({ effect: "deny", reason: "malformed_response" });
-        throw new KoiRuntimeError({
-          code: "PERMISSION",
-          message: `Malformed approval response for "${request.toolId}" — failing closed`,
-          retryable: false,
-        });
-      }
-
-      // Emit second audit entry and trajectory step for the approval outcome
-      const approvalDurationMs = clock() - approvalStartMs;
-      if (auditSink !== undefined) {
-        auditApprovalOutcome(
-          ctx,
-          resource,
-          approvalResult,
-          request.input,
-          approvalDurationMs,
-          auditSink,
-        );
-      }
-      emitApprovalStep(ctx, request.toolId, approvalResult, request.input, approvalStartMs);
-      stepEmitted = true;
-
-      if (approvalResult.kind === "deny") {
-        getTracker(ctx.session.sessionId as string).record({
-          toolId: resource,
-          reason: approvalResult.reason,
-          timestamp: clock(),
-          principal: ctx.session.agentId,
-          turnIndex: ctx.turnIndex,
-          source: "approval",
-        });
-
-        dispatchApprovalOutcome?.({ effect: "deny", reason: approvalResult.reason });
-        throw new KoiRuntimeError({
-          code: "PERMISSION",
-          message: `Tool "${request.toolId}" denied by approval handler: ${approvalResult.reason}`,
-          retryable: false,
-        });
-      }
-
-      // Handle "always-allow" — add to session's always-allowed set.
-      // Keyed by agentId + grantKey (exact-command hash) so an approval
-      // covers THIS command + argv only. A different argv of the same
-      // prefix (e.g. `git push --force`) must prompt again. For scope
-      // "always", also persist to durable storage (SQLite) under the
-      // same exact-command grant key.
-      if (approvalResult.kind === "always-allow") {
-        const sid = ctx.session.sessionId as string;
-        let allowed = alwaysAllowedBySession.get(sid);
-        if (allowed === undefined) {
-          allowed = new Set();
-          alwaysAllowedBySession.set(sid, allowed);
-        }
-        const sessionGrantKey = `${ctx.session.agentId}\0${grantKey}`;
-        allowed.add(sessionGrantKey);
-
-        // Persist to durable storage if scope is "always", store is configured,
-        // and a real user identity exists. Anonymous sessions cannot create durable
-        // grants — they silently downgrade to session scope.
-        // Fail-safe: if persist throws, the tool still executes (approval was given)
-        // but permanence is not recorded. The user gets re-prompted next session.
-        const grantUserId = ctx.session.userId;
-        const grantAgentId = persistentAgentId ?? ctx.session.agentId;
-        if (
-          approvalResult.scope === "always" &&
-          persistentStore !== undefined &&
-          grantUserId !== undefined
-        ) {
-          try {
-            persistentStore.grant(grantUserId, grantAgentId, grantKey, clock());
-          } catch {
-            // Approval was given — execute the tool. Permanence just wasn't recorded.
-          }
-        }
-
-        getTracker(sid).record({
-          toolId: resource,
-          reason: `always-allow granted (scope: ${approvalResult.scope})`,
-          timestamp: clock(),
-          principal: ctx.session.agentId,
-          turnIndex: ctx.turnIndex,
-          source: "approval",
-        });
-
-        // Dispatch before next() so the permission outcome is recorded even if the tool throws
-        dispatchApprovalOutcome?.({ effect: "allow" });
-        return next(request);
-      }
-
-      // Handle "modify" — use updated input
-      // Never cache modify results: the input rewrite is the safety mechanism,
-      // and caching would replay the original unsafe input on subsequent calls
-      if (approvalResult.kind === "modify") {
-        // Dispatch before next() so the permission outcome is recorded even if the tool throws
-        dispatchApprovalOutcome?.({ effect: "allow" });
-        return next({ ...request, input: approvalResult.updatedInput });
-      }
-
-      // Cache allow-only approvals (never modify — see above)
-      if (approvalCache !== undefined) {
-        const userId = ctx.session.userId ?? "__anonymous__";
-        const ctxStr = serializeTurnContext(ctx);
-        const cacheKey = computeApprovalCacheKey(
-          backendFingerprint,
-          ctx.session.sessionId as string,
-          userId,
-          ctx.session.agentId,
-          request.toolId,
-          request.input,
-          ctxStr,
-          request.metadata,
-          decision.reason,
-          grantKey,
-        );
-        if (cacheKey !== undefined) approvalCache.set(cacheKey);
-      }
-
-      // "allow" — dispatch before next() so outcome is recorded even if the tool throws
-      dispatchApprovalOutcome?.({ effect: "allow" });
-      return next(request);
-    } catch (e: unknown) {
-      // Emit a failure trajectory step for timeout/handler errors so they
-      // are observable in ATIF even though no valid decision was received.
-      // Skip if a step was already emitted (e.g., a deny that throws after emitting).
-      if (!stepEmitted) {
-        const reason =
-          e instanceof KoiRuntimeError && e.code === "TIMEOUT" ? "timeout" : "handler_error";
-        emitApprovalStep(
-          ctx,
-          request.toolId,
-          { kind: "deny", reason },
-          request.input,
-          approvalStartMs,
-        );
-      }
-      if (e instanceof KoiRuntimeError) throw e;
-      throw new KoiRuntimeError({
-        code: "INTERNAL",
-        message: `Approval handler error for "${request.toolId}"`,
-        retryable: false,
-        cause: e,
-      });
-    }
-  }
 }
